moved interface definitions to Interfaces added some file headers
