+++ src/Automation/CSE.Automation/Constants.cs
@@ -1,4 +1,7 @@
-﻿namespace CSE.Automation
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+namespace CSE.Automation
 {
     public sealed class Constants
     {
