+++ src/Automation/CSE.Automation/DataAccess/ActivityHistoryRepository.cs
@@ -1,8 +1,12 @@
-﻿using System;
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+
+using System;
 using System.Collections.Generic;
 using System.Configuration;
 using System.Text;
 using System.Threading.Tasks;
+using CSE.Automation.Interfaces;
 using CSE.Automation.Model;
 using Microsoft.Azure.Cosmos;
 using Microsoft.Azure.Cosmos.Table;
