--- src/Automation/CSE.Automation/DataAccess/IAuditRepository.cs
@@ -1,7 +1,0 @@
-﻿using CSE.Automation.Interfaces;
-using CSE.Automation.Model;
-
-namespace CSE.Automation.DataAccess
-{
-    internal interface IAuditRepository : ICosmosDBRepository<AuditEntry> { }
-}
