+++ src/Automation/CSE.Automation/Interfaces/IActivityHistoryRepository.cs
@@ -1,10 +1,9 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using CSE.Automation.Interfaces;
 using CSE.Automation.Model;
 
-namespace CSE.Automation.DataAccess
+namespace CSE.Automation.Interfaces
 {
     internal interface IActivityHistoryRepository : ICosmosDBRepository<ActivityHistory>
     {
