+++ src/Automation/CSE.Automation/Interfaces/IAuditRepository.cs
@@ -1,0 +1,6 @@
+﻿using CSE.Automation.Model;
+
+namespace CSE.Automation.Interfaces
+{
+    internal interface IAuditRepository : ICosmosDBRepository<AuditEntry> { }
+}
