+++ src/Automation/CSE.Automation/Services/AuditService.cs
@@ -2,7 +2,6 @@
 using System.ComponentModel;
 using System.Globalization;
 using System.Threading.Tasks;
-using CSE.Automation.DataAccess;
 using CSE.Automation.Extensions;
 using CSE.Automation.Interfaces;
 using CSE.Automation.Model;
